Random connector name generation
