+++ crates/pipeline-types/Cargo.toml
@@ -26,7 +26,6 @@ proptest-derive = { version = "0.3.0", optional = true }
 # Revert after https://github.com/paupino/rust-decimal/pull/637 is merged:
 rust_decimal = { git = "https://github.com/gz/rust-decimal.git", rev = "ea85fdf" }
 thiserror = "1.0"
-uuid = { version = "1.6.1", features = ["v7", "std", "serde"] }
 
 [dev-dependencies]
 lazy_static = "1.4.0"
