+++ crates/pipeline_manager/src/api/mod.rs
@@ -134,6 +134,7 @@ The program version is used internally by the compiler to know when to recompile
         crate::db::types::program::SqlCompilerMessage,
         crate::db::types::program::ProgramStatus,
         crate::db::types::program::ProgramConfig,
+        crate::db::types::program::ProgramInfo,
 
         // API key
         crate::db::types::api_key::ApiKeyId,
@@ -146,7 +147,6 @@ The program version is used internally by the compiler to know when to recompile
         crate::demo::Demo,
 
         // From the pipeline-types crate
-        pipeline_types::config::ProgramInfo,
         pipeline_types::config::PipelineConfig,
         pipeline_types::config::StorageConfig,
         pipeline_types::config::StorageCacheConfig,
