+++ crates/pipeline_manager/src/db/types/pipeline.rs
@@ -1,8 +1,8 @@
 use crate::db::error::DBError;
 use crate::db::types::common::Version;
-use crate::db::types::program::{ProgramConfig, ProgramStatus};
+use crate::db::types::program::{ProgramConfig, ProgramInfo, ProgramStatus};
 use chrono::{DateTime, Utc};
-use pipeline_types::config::{PipelineConfig, ProgramInfo, RuntimeConfig};
+use pipeline_types::config::{PipelineConfig, RuntimeConfig};
 use pipeline_types::error::ErrorResponse;
 use serde::{Deserialize, Serialize};
 use std::fmt;
